+++ src/RcppExports.cpp
@@ -522,6 +522,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// binary_lincbo_implications
+List binary_lincbo_implications(NumericMatrix I, bool verbose);
+RcppExport SEXP _fcaR_binary_lincbo_implications(SEXP ISEXP, SEXP verboseSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< NumericMatrix >::type I(ISEXP);
+    Rcpp::traits::input_parameter< bool >::type verbose(verboseSEXP);
+    rcpp_result_gen = Rcpp::wrap(binary_lincbo_implications(I, verbose));
+    return rcpp_result_gen;
+END_RCPP
+}
 // calculate_stability_sparse_rcpp
 NumericVector calculate_stability_sparse_rcpp(S4 mat);
 RcppExport SEXP _fcaR_calculate_stability_sparse_rcpp(SEXP matSEXP) {
@@ -990,6 +1002,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_fcaR_InClose_binary", (DL_FUNC) &_fcaR_InClose_binary, 3},
     {"_fcaR_calculate_grades_rcpp", (DL_FUNC) &_fcaR_calculate_grades_rcpp, 3},
     {"_fcaR_calculate_lattice_layout_rcpp", (DL_FUNC) &_fcaR_calculate_lattice_layout_rcpp, 5},
+    {"_fcaR_binary_lincbo_implications", (DL_FUNC) &_fcaR_binary_lincbo_implications, 2},
     {"_fcaR_calculate_stability_sparse_rcpp", (DL_FUNC) &_fcaR_calculate_stability_sparse_rcpp, 1},
     {"_fcaR_calculate_density_rcpp", (DL_FUNC) &_fcaR_calculate_density_rcpp, 3},
     {"_fcaR_calculate_separation_rcpp", (DL_FUNC) &_fcaR_calculate_separation_rcpp, 1},